Tools/pyproject.toml
[tool.poetry]
name = "stack-tools"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.2"
pillow = "^10.1.0"
pyperclip = "^1.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
